Fuzzily Search Help and View the Selection

Users no longer have to remember the exact name to find help, vignettes, and demo. A shiny gadget helps you to find a topic fuzzily. Click radio buttons to switch preview contents. Click "Done" or "Cancel" to close the widget. When background = FALSE, the "Done" button will also hook help, vignette, or demo, accordingly.

fuzzyhelp(
  query = "",
  method = getOption("fuzzyhelp.method", "fzf"),
  background = getOption("fuzzyhelp.background", TRUE),
  viewer = shiny::paneViewer()
)

Arguments

query

An initial query to search for the help system.

method

A fuzzy match method to use. Choices are "fzf" and "lv" (levenstein). The method "lv" is faster but can be less accurate. The default value can be tweaked by options(fuzzyhelp.method = "lv").

background

Whether to run a shiny gadget in a background process. The default value is TRUE and can be changed by option(fuzzyhelp.background = FALSE).

viewer

Specify where the gadget should be displayed--viewer pane, dialog window, or external browser--by passing in a call to one of the viewer() functions.

Value

If the background argument is TRUE, then the return value inherits from callr::r_bg(). Otherwise, NULL is returned.

Note

The default fuzzy match algorithm is a simplified version of https://github.com/junegunn/fzf. The implementation in this package excludes bonuses from relationship with matched characters and their previous characters.

Examples

if (FALSE) {
  fuzzyhelp()
}